fix value decoding
